change response process and done

+++ src/internal/domain/repository_interface/user_repository.rs
@@ -6,5 +6,5 @@ pub trait UserRepository {
     fn new() -> Self where Self: Sized;
     async fn create_name(&self, name: String) -> UserEntity;
     async fn get_names(&self) -> Vec<UserEntity>;
-    async fn delete_name(&self, index: usize) -> Result<(), String>;
+    async fn delete_name(&self, index: u32) -> Result<(), String>;
 }
